fix: :white_check_mark: designate officially supported versus unofficially supported transformers.
By default, only officially supported transformer types are tested in CI, but unofficially supported transformers MAY or MAY NOT be compatible with the `transformer_strategy`. A test fully parametrized over ALL of the model types in `transformers` is available, and can be enabled when setting the environment variable `HYPOTHESIS_TORCH_TEST_UNSUPPORTED_TRANSFORMERS=True`.